backend/requirements.txt
pypdf
langchain
langchain_community
langchain-google-genai
python-dotenv
bs4
langchain-text-splitters
pydantic
streamlit
langchain_chroma
requests
flask-cors
flask